split conversation and message, simplify message structure
